+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Release: dd.mm.yyyy
 ### Changes
 
 - Forward `console.log(...)` to `Phoenix.log(...)` for convenience.
+- Log stack trace when passing an error to `Phoenix.log(...)` ([#181](https://github.com/kasper/phoenix/issues/181)).
 
 2.6.7
 -----